+++ src/components/file/editForm/File.edit.display.js
@@ -3,4 +3,21 @@ export default [
     key: 'placeholder',
     ignore: true
   },
+  {
+    type: 'checkbox',
+    label: 'Files Synchronization feature',
+    tooltip: 'Enable ability to control files synchronization. Files will be auto synced before submit.',
+    key: 'autoSync',
+    input: true,
+    conditional: {
+      json: {
+        in: [
+          {
+            var: 'data.storage'
+          },
+          ['s3', 'azure', 'googledrive']
+        ],
+      }
+    }
+  },
 ];