+++ src/components/radio/fixtures/comp10.js
@@ -1,0 +1,21 @@
+export default {
+    'label': 'Radio',
+    'optionsLabelPosition': 'right',
+    'inline': false,
+    'tableView': false,
+    'values': [
+        {
+        'label': '01',
+        'value': '01',
+        'shortcut': ''
+        },
+        {
+        'label': '1',
+        'value': '1',
+        'shortcut': ''
+        }
+    ],
+    'key': 'radio',
+    'type': 'radio',
+    'input': true
+};